+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -252,9 +252,17 @@ class DirectivesChatbotPlug {
     }
     else if (directive_name === Directives.INTENT) {
       // console.log(".....DirIntent")
-      new DirIntent(context).execute(directive, async () => {
-        let next_dir = await this.nextDirective(this.directives);
-        this.process(next_dir);
+      new DirIntent(context).execute(directive, async (stop) => {
+        if (stop) {
+          if (context.log) { console.log("Stopping Actions on:", JSON.stringify(directive));}
+          this.theend();
+        }
+        else {
+          let next_dir = await this.nextDirective(this.directives);
+          this.process(next_dir);
+        }
+        // let next_dir = await this.nextDirective(this.directives);
+        // this.process(next_dir);
       });
     }
     else if (directive_name === Directives.MESSAGE) {