+++ tybotRoute/CHANGELOG.md
@@ -5,6 +5,18 @@
 available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
+### v0.2.15
+- replyto in send email fix
+
+### v0.2.14
+- pre-release
+
+### v0.2.13 - test
+- replyto send email test
+
+### 0.2.12
+- Updated @tiledesk/tiledesk-client => 0.10.4 to support the new email endpoint
+
 ### 0.2.11
 - TiledeskClient update to 0.10.3
 - updateLead() replaces the deprecated updateLeadData()